refactor(CLI): Drop the "transport" command group
I think this command group is probably now beyond the scope of alpenhorn.
Certainly it's beyond the scope of the main CLI.

All of the alpenhorn-specific functionality in these commands has
moved to the "node" group (`node init / activate / deactivate / list`).

So, remaining functionality in this command group is limited to:

  * partitioning disks
  * formatting partitions
  * mounting partitions
  * unmounting partitions

which are all very different that what alpenhorn usually does and
introduces unusual dependencies into alpenhorn.

I wouldn't be opposed to taking this functionality and combining it with
the higher-level commands that we used to have for CHIME in the
`alpenhorn_transport` script, but I don't plan to do that now and I
doubt anything will happen on that front without someone running an
alpenhorn transport fleet somewhere to motivate and provide a place for
testing.