fix(CLI): Re-implement sync command (#233)
There's now a --cancel flag that can be used to cancel pending
transfers. The long-deprecated --transport option has been removed. It
was needed before we implemented the --target option.

The --nice option is also removed. "Niceness" in ArchiveFileCopyRequests
was a concept that was never fully implemented and references to it have
been removed from the alpenhorn-2 daemon.

There's both "node sync NODE GROUP" and "group sync GROUP NODE". They're
almost interchangeable. The only differences are:
```
  group sync GROUP --cancel --all
  node sync NODE --cancel --all
```
which cancel all transfers into GROUP/out from NODE and can't be invoked
using the other command.

I have prohibited --target from being used with --cancel because I can't
figure out how it would be useful. If someone can come up with a
reasonable use case, that could be revisited.

Closes #106
Closes #172

Also, as this was the last top-level command to need re-implementation,
this also closes #122.