feat(CLI): node init command (#237)
Also a --init flag to "node create" to do the same thing.

This requests the daemon intialise the node, which typically means
creating the ALPENHORN_NODE file, but doesn't have to. The request is
made via the ArchiveFileImportRequest with path="ALPENHORN_NODE" (which
is otherwise a forbidden path to try to import).

This has involved some changes to the daemon to give it the ability to
create ALPENHORN_NODE files. The request can't be handled in the normal
part of the loop (which is skipped for uninitialised nodes), so instead
it's handled during the `check_init` test in the early part of the
update loop.

I've also taken the opportunity to disambiguate two different uses of
the term "active" which used to be in the daemon:

* the `StorageNode.active` bit is set
* the I/O layer is ready for I/O

The second of these is now called "initialised" and "active" now only
refers to the first one.

Closes #227