feat(util): Make `pretty_bytes` more forgiving
While it was pretty easy to avoid passing weird stuff to `pretty_bytes`
in ths daemon, that's harder to do in the CLI.

I've found I've been doing the same None check a lot before using
the function to deal with the potential for nulls in the database.

So, let's make the function a little more forgiving about the input.
Not sure if "??? B" is the best thing to return in the None case. Let me
know if you have a preference for something else.