feat(CLI): "file list" command
Usage here is a little complex.  Might need better help to explain which options can be used together.  Also: is `--not` a too confusing for the flag which negates the node/group constraint?

Also, the distinction between `--not` and `--not --all` might be a little subtle:
```
   file list --not --node=Node1 --node=Node2
```
lists files that are absent from both Node1 and Node2
```
   file list --not --all --node=Node1 --node=Node2
```
lists files which are absent from at least one of Node1 or Node2. Does that feel backwards?

Documentation with examples may help somewhat.

I found it difficult to collapse this command into a single query, so instead there are three stages (excluding the various name-to-thing look-ups):

1. one or more query to handle --node --group --all and state flag constraints
2. a query, maybe, for --from and --to
3. a final query that ties the above together with the --acq limit, and applies the --not, if necessary.

Closes #45 via
```
file list --corrupt --node=<NODE>
```