feat(CLI): "file create" command
Not the way you should be creating ArchiveFiles, but
provided for completeness (like "acq create" was).

Has some actual I/O in it: it can scan a file to generate a MD5 hash if
you want it to.  Other than that, this is another simple command: it
just creates an ArchiveFile record.