feat(CLI): "file list" command (#247)
Usage here is a little complex. Might need better help to explain which
options can be used together. Also: is `--not` a too confusing for the
flag which negates the node/group constraint?

I found it difficult to collapse this command into a single query, so
instead there are three stages (excluding the various name-to-thing
look-ups):

1. one or more query to handle location and state flag constraints
2. a query, maybe, for --from and --to
3. a final query that ties the above together with the --acq limit, and
applies the --not, if necessary.

Closes #45 via:
```
file list --corrupt --node=<NODE>
```