fix(import): Vet acq and file names.
This creates a function, `common.util.invalid_import_path`, that
takes an acq name, file name or both, and checks whether it
can be used.

This is then used in four places to try to reduce the number of
invalid acq/file names that make it into the database:

* in the "acq create" CLI command
* in the "file create" CLI command
* to check the acq-name returned by the import-detect extension
* when processing non-recursive ArchiveFileImportRequests

I have also added a check for recursive ArchiveFileImportRequests that
the requested scan path is resolvable and in-tree.

I think it's still fairly easy to mailicously add such names to
the database, but this should deal with inadvertantly providing
invalid names.