feat(CLI): "file find" command
After "file list", this was less involved that I feared it would be.

I think there's potential to add more ways of searching for files, but
this is good for now.  There are always going to be limits to how much
selection can be done with the CLI.  For very complex queries, users are
going to have to access the ORM directly and make their own queries.

This is the last command I wanted to implement from the command list.
So, I'm going to say this closes #202, unless someone can come up with
an argument to promote one of the proposed commands I suggested, which
otherwise will get left on the back-burner.

After all these command PRs are merged there is still going to be one or
more clean-up passes to do on the CLI before I'm willing to leave it
alone for the time-being.