fix(ui): Better error message for no config
The error raised when no config files can be found will now always be
ClickException and the error text now points the user to the new
--help-config option which will provide basic help on what needs to be
configured for things to work.  (And refer the user to the as-yet
non-existent documentation for full details.)

I was originally going to only do this in the CLI, but I think it's
helpful to people running the daemon to also see it there.

The --help-config option is an eager option, which means it can be used
without needing the full alpenhorn start-up sequence to complete.

Closes #245