fix(db): Better DB-connection failure handling (#263)
My goal here was to deal with #244 but that's led me to do some general
(and well-needed) clean-up of how the database connections are handled
in the test suite as well.

## Changes to the DB connection code

The first thing I've done is in `alpenhorn.db._base` where I've removed
the the fallback in-memory DB in the default `_connect` function which
used to be instantiated when no DB connection data was present in the
config. I think this was here to make testing alpenhorn easier, but it
causes trouble in production environments by hiding instances where the
config is not available.

Had we required all productions systems to use a database extension for
production database access, this might have been fine to leave as-is,
but in practice I don't see how we can do that.

Secondly, a connection error in `db.connect()` will now raise a
`click.ClickException` resulting in program exit. Before, adding this
what would usually happen was a crash later when first trying to use the
uninitialised database proxy.

## DB connection in the CLI

It appears I never actually had the CLI conencting to the database and
the test-suite ran because the fixtures were doing that on the side for
the client.

But now the connection is explicit. I've added the connection attempt to
the command group stubs, rather than the top-level CLI entry point,
which allows users to still do, say
```
  alpenhorn node --help
```
with an unconfigured database. I don't know if this is really necessary;
it was nice for looking at the usage output while developing, but it
might be better to complain as early as possible. Wherever we put it,
the user won't be able to do anything useful without a properly
configured database.

Closes #244 

## Test-suite fixes

These two changes have forced me to address the somewhat slap-dash way I
was creating and accessing test databases in the test-suite. The default
in-memory DB that was automatically created by an unconfigrued
`alpenhornd` is now made explicitly in the `dbproxy` fixture. Also, the
persistent in-memory database used by the client tests explicitly drop
all tables between tests to keep things isolated.