feat(daemon): Run-once mode
Adds an option -o/--once/--exit-after-update to the daemon
to make it run the update loop once, wait for all tasks to finish
(including deferred ones), and then exit cleanly.

The end-to-end test in `tests/daemon/test_entry.py` has been
updated to use this mode (rather than what we were doing before which
was using the `loop_once` fixture that co-opts the global abort).

Requires #268 to prevent the end-to-end test from taking ten minutes to
complete.