feat(io): Parameterise LustreHSM restore check time
This adds an optional key to the I/O config JSON for LustreHSM
nodes to allow specifying how long to wait between checks of in-progress
HSM restores.

I still think it would be nice to have some sort of global default
(see #197), but this will at give some ability to adjust the otherwise
hard-coded value.