fix(demo): Get the docker demo working again
This is a first pass at resuscitating and modernising
the docker-based demo.

It's not completely done; some of the demo needs restructuring for
modern alpenhorn, but basic functionality is restored, and it's
working up to around step 8 of `demo-script.md`.

One thing that I think will need changing is removing the `demo/data`
directory which the alpenhorn containers are mounting as a volume.
Running the demo results in `root`-owned files in the directory, which
isn't great.  I think it'll be possible to remove the `data` directory
completely and use normal (i.e. docker-managed) volumes instead, which
is likely what should be happening.

I've had to do some work on the `pattern_exporter.py` example extension,
which is used by the demo.  There's a new init function to populate the
tables for the demo, and a fix to the detection code.

The `demo/Dockerfile.alpenhorn` and `demo/alpenhorn.conf` files have been
resurrected from the test suite, where they had been deleted in #265.