doc(examples): Add example robot script
This is a generic version of the robot script used by CHIME on
cedar.  It's meant to provide some suggestions on how to run
alpenhornd non-interactively.

It probably needs some extra documentation to be build around it to
explain it more thorougly.

Also, I'm deleting the old `alpenhorn_ensure_running.sh` script.  It
served a similar purpose back in the day, but our understanding on
how to manage `screen` sessions has since improved and the way that
script does it shouldn't be done anymore.  Better for people to folow
the example of the robot script.