feat(daemon): Prometheus client (#277)
This adds a prometheus client to the daemon which will be started if the
port it should listen on is specified in the alpenhorn config.

I've added metrics here and there. I suspect some changes will be
necessary to what data is produced, but that can happen iteratively.
It's difficult to know how we actually want to instrument the process
without some testing. If you see any glaring omission, feel free to
suggest it. In addition to the explicit metrics I've defined, some basic
stats about the python process itself, e.g. CPU load, memory usage, &c.
are automatically produced by the prom client server.

Metrics themselves are represented by a new `Metrics` class which mostly
wraps `prometheus_client.Gauge` and `prometheus_client.Counter`. I've
tried to make a system where we should only have to define a given
metric once, even if we want to access it from various places. (The
biggest annoyance with defining metrics is the need to specify the
description, which is completely unneeded once the metric has been
defined once, but is still required for each `Counter` or `Gauge`
instance.)

For a very few metrics, I couldn't find a natural place in the existing
codebase to define them, so they've been put in the `metrics.by_name`
function from which they can be retrieved.

I tried, at least, to come up with a set of metrics that could replace
the metrics created by the `alpenmonitor` that creates metrics for
alpenhorn on wind. But I looked at that, and there's shockingly few
metrics that it's actually producing (it's mostly just the top three
panels in the Alpenhorn grafana page).

Note: all metrics that the daemon presents are about the running daemon.
It does not provide any of the metrics currently produced by the CHIME
`archive_stats` job (which summarises the data index by directly
inspecting the database). There's no plan to add them, but it would
probably be a good idea to migrate that script over to this repo as a
genericly useful script, or maybe as some sort of `alpenhorn db stats`
command.

The `prometheus_client` dependency is optional. If it can't be found at
runtime, most of the metric-handling in `common.metrics` code does
nothing. The rest of alpenhorn is agnostic as to whether the
`prometheus_client` module was available or not.

Closes #166