fix(auto_import): Improvements found via the demo (#279)
A few fixs for the auto-import code after running it through its paces
in the docker demo.

Fix race condtion in `_import_file`
-----------------------------------

In `_import_file`, we now use a full get-create-get idiom to avoid race
conditions, i.e.:

* try to get an existing record
* if that fails because the record doesn't exist, try to create the
record
* if that fails because the record (now) exists, then get the existing
record

Previously the second get wasn't there, but that can cause problems if
another thread/process creates the record between steps one and two.
This is how `peewee`'s own `get_or_create` helper function works, but we
can't use that here because the `register` flag means we sometimes stop
the attempt if step one fails.

More filtering in the `RegisterFile` handler
--------------------------------------------

In `RegisterFile`, which is used as the callback for filesystem events
triggered by the auto-import watchdog, we now do more filtering. Now
directories and lock files are ignored. Previously they would be passed
on to the `import_file` task which would eventually discard them for
being un-importable.

I've also deleted the `on_modified` handler, which would fire when a
file on disk was modified. I don't know when this would be useful in a
practical sense. I suppose it _could_ be leveraged by an intelligent
`import-detect` extension that defers registering a file until it has
certain contents, but I think that's a bad pattern to support.

Better to have these systems simply keep a lock-file around until
they're ready to have their file registered.

The `on_modified` handler also tended to result in duplicate jobs being
submitted for a newly created file because, on many filesystems, a file
creation (triggering `on_created`) is immediately followed by a metadata
update for the file which would also trigger `on_modified` for the same
newly created file.

If someone can come up with another reason why it would be useful to
have `on_modified`, though, I'd be willing to restore it.

Neither of these changes fundamentally change how alpenhorn was
importing files, because none of the things which have been removed
would have reuslted in importable files, but they both do reduce the
number of destined-to-fail import jobs being queued.