fix(tests): Fix mock_stat fixture for pyfakefs>=5.8
And by "fix" I mean delete.

This fixture was mocking `pathlib.PosixPath.stat` within the fake
filesytem.  But that method has worked, by default, within the fake
filesystem since `pyfakefs-5.4`.

And now `pyfakefs-5.8` has broken the implementation that we
were using in the `mock_stat` fixture.

So, the solution here is to delete the fixutre and require
`pyfakefs>=5.4`.