fix(tests): Fix mock_stat fixture for pyfakefs>=5.8 (#282)
And by "fix" I mean delete.

This fixture was mocking `pathlib.PosixPath.stat` within the fake
filesystem. But that method has worked, by default, within the fake
filesystem since `pyfakefs-5.4`.

And now `pyfakefs-5.8` has broken the implementation that we were using
in the `mock_stat` fixture.

So, the solution here is to delete the fixutre and require
`pyfakefs>=5.4`.

Also: since I'm messing about in the dependencies, I've dropped the
`docker` dependency from the test requirements: we don't use docker for
testing anymore.