perf(db): Re-implement StorageNode.get_all_files
This replaces the previous implementation, which ended up performing
one query per file present on the node, with a faster one which
invovles only a single query.

The old implementation worked fine in the test suite where there
might be, say, five files, but took tens of minutes on gong, where
there are ~10,000 files.  This new implementation takes seconds on
gong.