feat(common.config): Test isolation (#286)
This is not the cause of the test failures in
https://github.com/chime-experiment/alpenhorn-chime/pull/5, but while
debugging those failures on `wind` I discovered that installed
`alpenhorn.conf` files will mess up those tests.

So we need a way to tell alpenhorn to ignore installed config files so
that tests can run in isolation.

Adds a `--test-isolation` flag to both CLI and daemon that will cause
alpenhorn to skip loading config from the standard paths (envar and
--conf still work). And also a function in `common.config` to accomplish
the same.

Note: this is not needed in the alpenhorn test suite itself, because
everything in the test suite that looks for config files runs in a fake
filesystem.