feat(cli): Read file list via --from-file
This adds a `--from-file` option to various commands to allow
limiting the command to a list of ArchiveFiles specified in a text
file.

The intention is that, say,
```
        alpenhorn node sync cedar_nearline cedar_online --from-file=<FILE>
```
should do the same thing as
```
        recall_filelist <FILE>
```
is currently doing on cedar.

Can also read from stdin (via `--from-file=-`), in which case `--check`
mode will be automatically turned on if the user doesn't invoke `--force`.
(Because the file-reading code will consume all the input, making it
impossible for the user to interactively confirm the operation, in that
case.)

The commands which accept this option are:

* `node clean`
* `node sync` (AKA `group sync`)
* `node verify`

Note: `file create` also has a `--from-file` option, but that one is
completely different (it's a boolean flag there).

Also: renamed the `--count` flag in `node verify` to `--check` to make
it the same as all the other commands with a check/confirm/update flow.
Not sure why it was different; the `node verify` tests seem to indicate I
had been assuming this rename already.