feat(cli): Read file list via command-line option (#287)
This adds a `--file-list` option to various commands to allow limiting
the command to a list of ArchiveFiles specified in a text file.

The intention is that, say,
```
        alpenhorn node sync cedar_nearline cedar_online --file-list=<FILE>
```
should do the same thing as
```
        recall_filelist <FILE>
```
is currently doing on cedar.

Can also read from stdin (via `--file-list=-`), in which case `--check`
mode will be automatically turned on if the user doesn't invoke
`--force`. (Because the file-reading code will consume all the input,
making it impossible for the user to interactively confirm the
operation, in that case.)

The commands which accept this option are:

* `node clean`
* `node sync` (AKA `group sync`)
* `node verify`

Also: renamed the `--count` flag in `node verify` to `--check` to make
it the same as all the other commands with a check/confirm/update flow.
Not sure why it was different; the `node verify` tests seem to indicate
I had been assuming this rename already.

Closes #285