feat(auto_import): Reject files with leading dots
While working on the demo script, I realised it's very tricky to
talk about creating files on a StorageNode which has auto-import
turned on which aren't going to be imported, because, by default,
`alpenhorn` has no criteria to reject certain filenames.  I've been
mulling this over, and I think this is likely a good way of dealing with
this idea.

This changes the daemon import to refuse to register any file
whose filename starts with a leading dot.  While this does put a
mild restriction on users' ability to name their data files, the benefit
is there is always a portion of the namespace that can be used without
worring about running afoul of alpenhorn.

As a side benefit, this removes the confusion over what a lockfile
for a file with a leading dot would be;  i.e. would the lockfile
for a file named `.name.h5` be `.name.h5.lock` or `..name.h5.lock`?
(Answer: it's the second one.)

Note: this is only _filenames_ which may not have a leading dot;
directory names, including one used as an `ArchiveAcq.name` may
still start with a leading dot.

There are two parts to this change:
* A check in `_import_file` will reject any file with a leading
  dot in its name.
* A change to the RegisterFile watchdog event handler which means
  events for files with leading dots are no longer forwarded on to the
  import machinery.