build: delete install directory (#291)
Not only is this stuff way out-of-date and too CHIME-specific, I also
don't like having a file here called `alpenhornd.conf` that isn't the
alpenhorn config file.

In place of this, I've added an example systemd service file (which is
not much more than a copy-paste from the one on `wind`) to the examples
directory.