fix(cli): Handle negative sizes in DB
The alpenhorn-1 way to indicate no maximum size for a node was
to set `StorageNode.max_total_gb` to -1.  Alpenhorn-2 made things
more intuitive by using Null/None in this case, but we still need
to be able to handle negative sizes from the database.

Other than `max_total_gb`, there's no reason to expect the other `..._gb`
fields in the database to have a negative value, but they're all just
floats, so there's nothing to prevent it, so I've added a check for all
the fields.