test: Ignore ALPENHORN_CONFIG_FILE in pytest's environment
If the execution environment for `pytest` contains
`ALPENHORN_CONFIG_FILE`, we want to ignore that while testing (because
such a file probably isn't going to be available in the fake filesytem).

I've added this fix to the `xfs` fixture, since it's never going to be a
valid envar when the fake filesystem is in use, but it could probably go
elsewhere in the fixtures, if we can figure out a better place.