feat(cli): Allow node.root in file-list
This is an attempt to make file-list parsing a little more forgiving.
Removes `node.root` from files listed in the file-list, if that is
present.

For transfers, this will only allow the source `node.root` (because
the destination node can't be determined by the CLI).  In cases where
no node is given (e.g. `group sync --cancel --all --file-list=...`)
no such path removal is attempted.