feat(db): Implement schema versioning (#300)
Adds schema versioning to the database. There's no support for schema
migration (that may have to come later, if we ever change the database
schema), but I think it's useful to at least set up schema version
tracking before we release alpenhorn more generally. This way, if we
_do_ decide to change the database schema later at the very least users
will be able to tell whether the alpenhorn they're using is compatible
with the database they have.

Adds a new table, `DataIndexVersion`, with columns as outlined in #62 to
track schema versioning (which can be used by extensions, if they so
desire). Also adds a new global `alpenhorn.db.current_version`
containing the currently-supported schema version, which I've set to 2,
leaving 1 to represent all old ("CHIME-style") data index versions.

Both the daemon and the CLI will now check the schema version reported
by the database on start-up and refuse to run if it's different than the
version they expect. The sole exception here is the `alpenhorn db`
command group, which skips the check, to allow using this group when the
database hasn't been initialised yet. However, the `db init` command now
will check whether the databse already exists (even in part) and refuse
to operate if it does. The sole exception being: if the
database-reported schema version is equal to the version `db init` was
going to create, then it successfully does nothing.

To avoid constantly having to update it all over the place, I've also
created a `alpenhorn.db.gamut` tuple which lists all tables in the data
index, meaning there's only one place the whole list needs to be
maintained now.

Closes #62
Closes #271

NB: this breaks the demo script as written because, with this change,
initialising the pattern importer before creating the data index will
cause `db init` to detect a broken data index and refuse to complete.
This is my secret motivation behind #299.

Also: before this is deployed, we will need to add `DataIndexVersion` to
the CHIME data index so the new checks don't stop the daemon and CLI
from running.