refactor(common.util): Remove unused `hr` param from `md5sum_file` (#304)
Alpenhorn has never set `hr` to anything other than it's default (True).

It's not clear why this parameter is here, but it's a clear hold-over
from alpenhorn-1, where the function may have been meant to be used as a
general-purpose utility function. This function was never a part of the
alpenhorn-1 code base per se, it was part of the mess of
alpenhorn-adjacent code that lived in `chimedb.data_index.util`, where
it also had a `cmd_line` flag to use the `md5sum` binary, instead of
in-Python computation (in which case, the `hr` flag went ignored).

cf. https://github.com/chime-experiment/chimedb_di/pull/28