Mypy / ruff format fixes (#303)
## Errors found by `mypy`

`ruff`, explicitly, doesn't do type checking with type hinting, so I've
experimented a bit with running `mypy` on the code to see what it would
say. The primary result is now I have Opinions on `mypy`, which this PR
is too narrow to contain. But has also found some problems that I've
fixed here, which is great. Still, I'm not convinced it's something we
want running in our CI...

Most of these fixes are fixes to the type hints themselves, which is
what I was expecting `mypy` to help me with, but there are a few actual
(albeit minor) bugs which have also been fixed.

## ruff format

I also experimented in this branch with `ruff format` as a replacement
for `black`. It produces results that are very similar (by design: the
`ruff` devs are overtly trying to mimic black), but it's not identical.
The thing I liked most about `ruff format` is its ability to concatenate
strings when the result would fit on a single line.

The code here has been run through `black` and then `ruff format` and
then back through `black`, meaning there are a few formatting changes in
the code that were changed by `ruff` from what `black` had done, but
then were considered acceptable by the second `black` run, and not
reverted. (Several other changes made by `ruff format` _were_ reverted.)

I think the formatting done by `ruff format` is good enough that, were
we using `ruff` everywhere I'd suggest using it for formatting as well
as linting, and ditch `black`, just to minimise the number of tools we
have in play. But, since we're not using `ruff` everywhere, I don't
think it's worth switching. We definitely shouldn't get ourselves into a
situation where were using `black` in some projects and `ruff format` in
others.