feat(scheduler): Implement FIFO clearing
Adds a method, `clear_fifo`, to the queue which will dequeue all
pending and deferred tasks in the specified FIFO.  After doing this,
the FIFO may be optionally closed, preventing further putting new
items into it.  Calling `put` on a closed FIFO will raise `KeyError`.

In-progress tasks are not affected, though if the FIFO they're
running out of is closed, and these tasks requeue or defer themselves,
they won't successfully be able to do that.

The `Task` has been updated to be able to handle this `KeyError` (by
doing nothing other than printing a message that the item was
discarded).  However, nothing (yet) invokes `clear_fifo`, so this
situation will not happen.

Also: moved the `task_done` method without change to better organise
the queue module.