fix(daemon): Dequeue tasks when stopping updates.
When a node stops being updated (e.g. when deactivated), now pending
tasks for the node are discarded from the queue.  This will happen
both when the node/group is no longer available (because it was
deactivated or the filesystem went away), but also as part of a
re-init, since that's going to restart the I/O framework for the
storage node/group.

This is effectively the same as killing the daemon and restarting it,
(but only for this node), but it turns out that's a fairly safe thing
to do, because we've done a good bit of work to make the daemon robust
against that situation over the years. So, despite my original opinion on
the option,  this closes #309 using the first of the proposed solutions.

I think this is the right way to deal with this because it also fixes
another problem which I've been mulling over: how to deal with pending I/O
when a node goes away.  And this seems like the right solution to that
problem too: i.e. just throw it all away, and assume some future daemon
will figure it out.