doc: Update demo per Josh's notes (#316)
_Read this version of the demo on RTD here:
https://alpenhorn.readthedocs.io/en/doctest/demo.html_

An unfortunately large fraction of the changes here are simply RST
formatting fixes (and adding things like hint and tip callouts).

Important changes:
* I've rotated the background highlight for stuff occurring within the
docker containers by 180 degrees in hue (meaning the background is a
not-entirely-unpleasant purplish instead of yellow) with the hope that
it makes it more clear where specific actions take place.
* I've added an extra container to the demo, which I've called
`alpenshell` where the demo suggests you run the `alpenhorn` CLI
commands (though it does point out you could use one of the storage
hosts, too). My reasoning is twofold:
* it reinforces the idea that the CLI can be run anywhere that can
access the database, and does not need to be run on a node containing a
daemon or StorageNode
* it prevents the awkwardness of having to start up a temporary version
of `alpenhost1` to bootstrap the demo.
* the downside to the above is there's a potential for confusion when
doing manual file creation on `alpenhost1`, since that can't be done on
`alpenshell` (because it doesn't have the `demo_storage1` filesystem).
I've tried to guard against this by throwing in some `cd /data` commands
into the `alpenhost1` command lists and given it its own highlight colour.

Individual comments are addressed below.