fix(StorageNode): Include suspect files in size count.
This adds files with `has_file='M'` to the total size for a node calculated
by `StorageNode.get_total_gb`.

A downstream daemon pulling files can't always tell why transfers fail.
As a result, in times of network trouble, it's not unusual for a
pulling daemon to mark files it fails to transfer as suspect (`has_file="M"`).
In these cases, because the file is okay and it was the network that was
the problem, the upstream deamon will check the file and decide it's
okay, setting `has_file` back to "Y".  The net effect is a downwards
transient in the reported size of a node until the networking trouble
gets resolved, which can have follow-on effects such as over-filling a
node.

Because it triggers a check of the file, the `has_file="M"` state is
temporary (assuming a node is being managed).  So this shouldn't change
the reported size of anything in the long-term.